+++ CHANGELOG.rst
@@ -14,8 +14,9 @@ Added
 -----
 - `HTML reports`_ now have an "options" menu
   to filter exchanges and notices on the fly.
-- The ``httpolice`` command-line tool now exits with a non-zero status
-  if it has found any errors in the input data.
+- The ``httpolice`` command-line tool now has
+  a ``--fail-on`` option to exit with a non-zero status
+  if any notices with a given severity have been reported.
 - Check for missing scheme name in authorization headers (notice `1274`_).
 - Check for missing quality values in headers like Accept (notice `1276`_).
 - Check for obsolete 'X-' prefix in experimental headers (notice `1277`_).